SistemaImobiliaria/requirements.txt
astroid==2.0.1
certifi==2019.3.9
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.5.0
Django==2.0.7
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
model-mommy==1.6.0
Pillow==5.3.0
psycopg2==2.7.5
pylint==2.0.0
pytz==2018.5
requests==2.22.0
six==1.11.0
typed-ast==1.3.5
urllib3==1.25.2
virtualenv==16.5.0
whitenoise==3.3.1
wrapt==1.10.11
